font change, game class created, phrase class created
I added a google api font called Press Start 2P to give my game an 80s feel.  I also created two classes called game class and phrase class.  The phrase class creates a phrase object to be using in the game.  The game class handles the game parameters and the initialization of phrases objects to be used in the game.