bug fix, handleInteration function, minor fixes and resetGame function
I realized that when you would highlight the empty gray phrase the word was revealed.  To fix this a change was made to the ul li tags in the CSS file.

The handle Interaction function now works as intended when a user selects a button.

A resetGame function was created to reset the game when the user selects the start game button again after playing a game.